ocr/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30
pydantic>=2.6,<3

requests>=2.31
httpx>=0.27

# OCR stack
kraken>=6.0,<7
torch>=2.2,<3
pillow>=10.0
numpy>=1.26
scikit-image>=0.22
scipy>=1.10
PyMuPDF>=1.24.10  # provides `import fitz`

# tooling / tests
pytest>=8.3
ruff>=0.6.9

# document conversion
pdfminer.six>=20240706
python-docx>=1.1.2
lxml>=5.3
